Reorganize ComplexDoubleArray by protocol
