--- Sources/VecLab/Types/ComplexArray/ComplexDoubleArray+Collection.swift
@@ -1,7 +1,0 @@
-//
-//  ComplexDoubleArray+Collection.swift
-//  VecLab
-//
-//  Created by Marcus Painter on 11/05/2025.
-//
-
