+++ Sources/VecLab/Types/ComplexArray/Protocols/ComplexDoubleArray+Hashable.swift
@@ -1,0 +1,15 @@
+//
+//  ComplexDoubleArray+Hashable.swift
+//  VecLab
+//
+//  Created by Marcus Painter on 11/05/2025.
+//
+
+extension ComplexDoubleArray: Hashable {
+    
+    // MARK: - Hashable
+    public func hash(into hasher: inout Hasher) {
+        hasher.combine(real)
+        hasher.combine(imag)
+    }
+}
